offersliece rename dirrectory
