+++ src/store/offer-slice/offer-slice.ts
